Require higher level permissions for agile changes

+++ modules/agile/controllers/Main.php
@@ -124,12 +124,24 @@
 
             if ($request->isDelete())
             {
+            	// Check for scrum permission to modify boards
+            	if (!$this->_checkProjectPageAccess('project_scrum'))
+            	{
+            		throw new \Exception($this->getI18n()->__("You don't have access to modify boards (scrum permission)"));
+            	}
+            	
                 $board_id = $this->board->getID();
                 $this->board->delete();
                 return $this->renderJSON(array('message' => $this->getI18n()->__('The board has been deleted'), 'board_id' => $board_id));
             }
             elseif ($request->isPost())
             {
+            	// Check for scrum permission to modify boards
+            	if (!$this->_checkProjectPageAccess('project_scrum'))
+            	{
+            		throw new \Exception($this->getI18n()->__("You don't have access to modify boards (scrum permission)"));
+            	}
+            	
                 $this->board->setName($request['name']);
                 $this->board->setDescription($request['description']);
                 $this->board->setType($request['type']);
@@ -339,6 +351,12 @@
             {
                 if ($request->isPost())
                 {
+                	// Check for scrum permission to modify boards
+                	if (!$this->_checkProjectPageAccess('project_scrum'))
+                	{
+                		throw new \Exception($this->getI18n()->__("You don't have access to modify boards (scrum permission)"));
+                	}
+                	
                     $columns = $request['columns'];
                     $saved_columns = array();
                     $cc = 1;
